[pc-49] Link session to certification candidate
