+++ api/tests/acceptance/application/session/session-controller-get-attendance-sheet_test.js
@@ -1,0 +1,68 @@
+const { expect, databaseBuilder, generateValidRequestAuhorizationHeader } = require('../../../test-helper');
+const createServer = require('../../../../server');
+
+describe('Acceptance | Controller | session-controller-get-attendance-sheet', () => {
+
+  let server;
+
+  beforeEach(async () => {
+    server = await createServer();
+  });
+
+  describe('GET /api/sessions/{id}/attendance-sheet', () => {
+    let user, sessionIdAllowed, sessionIdNotAllowed;
+    beforeEach(async () => {
+      // given
+      user = databaseBuilder.factory.buildUser();
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      databaseBuilder.factory.buildCertificationCenterMembership({ userId: user.id, certificationCenterId });
+
+      const otherUserId = databaseBuilder.factory.buildUser().id;
+      const otherCertificationCenterId = databaseBuilder.factory.buildCertificationCenter().id;
+      databaseBuilder.factory.buildCertificationCenterMembership({ userId: otherUserId, certificationCenterId: otherCertificationCenterId });
+
+      sessionIdAllowed = databaseBuilder.factory.buildSession({ certificationCenterId }).id;
+      sessionIdNotAllowed = databaseBuilder.factory.buildSession({ certificationCenterId: otherCertificationCenterId });
+
+      await databaseBuilder.commit();
+    });
+    afterEach(() => databaseBuilder.clean());
+
+    it('should respond with a 200 when session can be found', async () => {
+      // when
+      const authHeader = generateValidRequestAuhorizationHeader(user.id);
+      const token = authHeader.replace('Bearer ', '');
+      const options = {
+        method: 'GET',
+        url: `/api/sessions/${sessionIdAllowed}/attendance-sheet?accessToken=${token}`,
+        payload: {}
+      };
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(200);
+      });
+    });
+
+    it('should respond with a 403 when user cant access the session', async () => {
+      // when
+      const authHeader = generateValidRequestAuhorizationHeader(user.id);
+      const token = authHeader.replace('Bearer ', '');
+      const options = {
+        method: 'GET',
+        url: `/api/sessions/${sessionIdNotAllowed}/attendance-sheet?accessToken=${token}`,
+        payload: {}
+      };
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(403);
+      });
+    });
+  });
+
+});
