+++ api/tests/acceptance/application/session/session-controller-patch_test.js
@@ -1,0 +1,117 @@
+const { expect, databaseBuilder, generateValidRequestAuhorizationHeader } = require('../../../test-helper');
+const createServer = require('../../../../server');
+
+describe('Acceptance | Controller | session-controller-path', () => {
+
+  let server;
+
+  beforeEach(async () => {
+    server = await createServer();
+  });
+
+  describe('PATCH /api/sessions/{id}', () => {
+
+    let user, unauthorizedUser, certificationCenter, session, payload;
+
+    beforeEach(async () => {
+      user = databaseBuilder.factory.buildUser();
+      unauthorizedUser = databaseBuilder.factory.buildUser();
+      certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+      databaseBuilder.factory.buildCertificationCenterMembership({
+        userId: user.id,
+        certificationCenterId: certificationCenter.id
+      });
+      session = databaseBuilder.factory.buildSession({
+        certificationCenter: certificationCenter.name,
+        certificationCenterId: certificationCenter.id,
+        address: 'Nice',
+        room: '28D',
+        examiner: 'Antoine Toutvenant',
+        date: '2017-12-08',
+        time: '14:30',
+        description: 'ahah',
+        accessCode: 'ABCD12'
+      });
+      databaseBuilder.factory.buildCertificationCourse({
+        sessionId: session.id
+      });
+      payload = {
+        data: {
+          id: session.id,
+          type: 'sessions',
+          attributes: {
+            address: 'New address',
+            room: 'New room',
+            examiner: 'Antoine Toutvenant',
+            date: '08/12/2017',
+            time: '14:30',
+            description: 'ahah',
+            accessCode: 'ABCD12'
+          }
+        }
+      };
+
+      await databaseBuilder.commit();
+    });
+
+    afterEach(() => {
+      return databaseBuilder.clean();
+    });
+
+    it('should respond with a 200 and update the session', function() {
+      const options = {
+        method: 'PATCH',
+        url: `/api/sessions/${session.id}`,
+        headers: { authorization: generateValidRequestAuhorizationHeader(user.id) },
+        payload
+      };
+
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise
+        .then((response) => {
+          expect(response.statusCode).to.equal(200);
+          expect(response.result.data.type).to.equal('sessions');
+          expect(response.result.data.id).to.equal(session.id.toString());
+          expect(response.result.data.attributes.address).to.equal('New address');
+          expect(response.result.data.attributes.room).to.equal('New room');
+        });
+    });
+
+    it('should return 404 HTTP status code when session cannot be found', () => {
+      // when
+      const promise = server.inject({
+        method: 'PATCH',
+        url: '/api/sessions/2',
+        headers: { authorization: generateValidRequestAuhorizationHeader(user.id) },
+        payload
+      });
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(404);
+      });
+    });
+
+    it('should respond with a 403 when user is not authorized to update the session', function() {
+      const options = {
+        method: 'PATCH',
+        url: `/api/sessions/${session.id}`,
+        headers: { authorization: generateValidRequestAuhorizationHeader(unauthorizedUser.id) },
+        payload
+      };
+
+      // when
+      const promise = server.inject(options);
+
+      // then
+      return promise.then((response) => {
+        expect(response.statusCode).to.equal(403);
+      });
+    });
+
+  });
+
+});
